# docs/openapi.yml
openapi: 3.0.3
info:
  title: Marketplace API
  description: >-
    Упрощенная система маркетплейса. Включает сервис продуктов и сервис заказов.
    Реализовано на стеке: ASP.NET Core, Dapper, PostgreSQL.
    Все идентификаторы используют формат UUID (Guid).
    Для защищенных эндпоинтов требуется JWT токен в заголовке Authorization.
  version: 1.0.0
  contact:
    name: Команда разработки
    email: team@marketplace.local
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Локальный сервер разработки
  - url: https://api.marketplace.example.com
    description: Продакшен-сервер
...
security:
  - JwtBearer: []
paths:
...
  /api/orders:
    get:
      tags:
        - Orders
      summary: Получение списка заказов текущего пользователя
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
          description: Номер страницы (начинается с 1)
        - in: query
          name: pageSize
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
          description: Количество элементов на странице
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/OrderStatus'
          description: Фильтр по статусу заказа
      responses:
        '200':
          description: Список заказов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderPagedResult'
        '401':
          description: Неавторизован

    post:
      tags:
        - Orders
      summary: Оформление нового заказа
      description: Создает заказ на основе переданного списка товаров из корзины клиента.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
            example:
              customerName: "Иван Иванов"
              customerEmail: "ivan@example.com"
              deliveryAddress: "г. Екатеринбург, ул. Ленина, 1"
              items:
                - productId: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                  quantity: 2
      responses:
        '202':
          description: Заказ принят в обработку (асинхронно)
          content:
            application/json:
              schema:
                type: object
                required: [orderId, statusCheckUrl]
                properties:
                  orderId:
                    type: string
                    format: uuid
                  status:
                    type: string
                    enum: [Draft, Processing]
                    default: Draft
                  statusCheckUrl:
                    type: string
                    format: uri
                    example: "/api/orders/{orderId}/status"
        '400':
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Неавторизован
        '409':
          description: Конфликт (товар недоступен)
  /api/orders/{id}:
    get:
      tags:
        - Orders
      summary: Получение информации о заказе
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Информация о заказе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDto'
        '401':
          description: Неавторизован
        '403':
          description: Доступ запрещен (заказ принадлежит другому пользователю)
        '404':
          description: Заказ не найден

    delete:
      tags:
        - Orders
      summary: Отмена заказа пользователем
      description: Позволяет пользователю отменить свой заказ, если он ещё не перешёл в статус "Собирается".
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Заказ отменен
        '400':
          description: Нельзя отменить заказ в текущем статусе
        '401':
          description: Неавторизован
        '403':
          description: Доступ запрещен
        '404':
          description: Заказ не найден

    put:
      tags:
        - Orders
      summary: Обновление статуса заказа (только администратор)
      security:
        - JwtBearer: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderStatusRequest'
      responses:
        '200':
          description: Статус обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDto'
        '400':
          description: Недопустимый переход статуса
        '401':
          description: Неавторизован
        '403':
          description: Недостаточно прав
        '404':
          description: Заказ не найден
  /api/orders/{id}/status:
    get:
      summary: Проверка статуса обработки заказа (для SAGA)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Текущий статус обработки
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { $ref: '#/components/schemas/OrderStatus' }
                  message: 
                    type: string
                    example: "Waiting for stock reservation..."
                  updatedAt: { type: string, format: date-time }
components:
  securitySchemes:
    JwtBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Введите JWT токен без префикса "Bearer ".
        Пример: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
  schemas:
...
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: string
          pattern: '^\d+\.\d{2}$'
          example: "1299.99"
          description: Сумма в формате строки для сохранения точности (2 знака после запятой)
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
          default: RUB
          example: "RUB"
          description: Код валюты по стандарту ISO 4217 (например, "RUB", "USD", "EUR", "KZT")
...
    CreateOrderRequest:
      type: object
      required:
        - items
      properties:
        customerName:
          type: string
          minLength: 2
          maxLength: 100
          description: "Имя получателя. Если не указано, берётся из профиля пользователя"
          example: "Иван Иванов"
        customerEmail:
          type: string
          format: email
          description: "Email для уведомлений. Если не указан, берётся из профиля"
          example: "ivan@example.com"
        items:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: '#/components/schemas/OrderItemCreate'
        deliveryAddress:
          type: string
          maxLength: 500
          nullable: true
          example: "г. Екатеринбург, ул. Ленина, 1, кв. 42"
    OrderItemCreate:
      type: object
      required:
        - productId
        - quantity
      properties:
        productId:
          type: string
          format: uuid
          description: Уникальный идентификатор товара (UUID)
          example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
        quantity:
          type: integer
          minimum: 1
          maximum: 99
          example: 2
    OrderDto:
      type: object
      required:
        - id
        - status
        - createdAt
        - totalAmount
        - items
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
          example: "7c9e6679-7425-40de-944b-e07fc1f90ae7"
        status:
          $ref: '#/components/schemas/OrderStatus'
          readOnly: true
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        customerName:
          type: string
          readOnly: true
        customerEmail:
          type: string
          format: email
          readOnly: true
        deliveryAddress:
          type: string
          readOnly: true
          nullable: true
        totalAmount:
          $ref: '#/components/schemas/Money'
          readOnly: true
          description: Итоговая сумма заказа
        items:
          type: array
          readOnly: true
          items:
            $ref: '#/components/schemas/OrderItemDto'
    OrderItemDto:
      type: object
      required:
        - productId
        - productName
        - quantity
        - priceAtPurchase
      properties:
        productId:
          type: string
          format: uuid
          readOnly: true
        productName:
          type: string
          readOnly: true
        quantity:
          type: integer
          readOnly: true
          minimum: 1
        priceAtPurchase:
          $ref: '#/components/schemas/Money'
          readOnly: true
          description: Цена товара на момент покупки
    OrderPagedResult:
      type: object
      required:
        - items
        - totalCount
        - page
        - pageSize
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderDto'
        totalCount:
          type: integer
          minimum: 0
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 100
    UpdateOrderStatusRequest:
      type: object
      required:
        - newStatus
      properties:
        newStatus:
          $ref: '#/components/schemas/OrderStatus'
        comment:
          type: string
          maxLength: 500
          nullable: true
          description: Комментарий к смене статуса (опционально)
    OrderStatus:
      type: string
      enum:
        - Created
        - Paid
        - Assembling
        - Shipping
        - Delivered
        - Cancelled
      description: Жизненный цикл заказа
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          format: uri
          example: "https://httpstatuses.com/400"
        title:
          type: string
          example: "One or more validation errors occurred."
        status:
          type: integer
          minimum: 400
          maximum: 599
          example: 400
        detail:
          type: string
          nullable: true
          example: "The field 'email' must be a valid email address."
        instance:
          type: string
          format: uri
          nullable: true
          example: "/api/orders"
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            email: ["Неверный формат email", "Email уже зарегистрирован"]
...